Fix crash in the Status method
As it turns out, hbsdctrl_state_to_string does not allocate the
string it returns on the heap and thinking that it did was a
critical mistake that led to a program crash.

At the same time, we have added a test that can detect the crash in
case it is ever introduced again (hopefully not).